FIX bug: Edit Ingredient -> nutritional value float - int parsing

+++ src/main/java/cz/muni/fi/pv168/project/ui/dialog/EditIngredientDialog.java
@@ -22,7 +22,7 @@ public class EditIngredientDialog extends EntityDialog<Ingredient> {
 
     private void setValues() {
         name.setText(ingredient.getName());
-        nutritionalValue.setText(Float.toString(ingredient.getCaloriesPerUnit()));
+        nutritionalValue.setText(Integer.toString(ingredient.getCaloriesPerUnit()));
         baseUnit.setModel(new javax.swing.DefaultComboBoxModel<Unit>(unitTableModel.getUnits().toArray(new Unit[0])));
         baseUnit.getModel().setSelectedItem(ingredient.getBaseUnit());
     }
